Function smtp_message::next_crlf

pub fn next_crlf(buf: &[u8], state: &mut NextCrLfState) -> Option<usize>

Returns the index of the \n in the first \r\n of buf, or None if there was none yet. This will update state, the first call should pass in NextCrLfState::Start, and subsequent calls (until a non-None value is found) should just keep using the same reference.