Struct smtp_message::EscapedDataReader

pub struct EscapedDataReader<'a, R> { /* fields omitted */ }

AsyncRead instance that returns an unescaped DATA stream.

Note that:

If a line (as defined by b“\r\n“ endings) starts with a b’.’, it is an “escaping” dot that is not part of the actual contents of the line.

If a line is exactly b“.\r\n“, it is the last line of the stream this stream will give. It is not part of the actual contents of the message.

Implementations

impl<'a, R> EscapedDataReader<'a, R> where
R: AsyncRead,

pub fn new(buf: &'a mut [u8], unhandled: Range<usize>, read: R) -> Self

pub fn is_finished(&self) -> bool

Returns true iff the message has been successfully streamed to completion

pub fn complete(&mut self)

Asserts that the full message has been read (ie. .is_finished() would return true), then marks this reader as complete.

Note that this should be called before saving the stream, given that until .is_finished() has returned true it’s not yet sure whether the stream ended due to connection loss or thanks to the end of data marker being reached.

pub fn get_unhandled(&self) -> Option<Range<usize>>

Returns the range of data in the buf passed to new that contains data that hasn’t been handled yet (ie. what followed the end-of-data marker) if complete() has been called, and None otherwise.